[move] Diffをgenericパッケージに移す
